Supported LiquiBase instruction for SQLite databases: addDefaultValue, dropDefaultValue, dropPrimaryKey, modifyColumn, addUniqueContraint, dropUniqueContraint, dropNotNullConstraint.
git-svn-id: http://liquibase.jira.com/svn/CORE/trunk@688 e6edf6fb-f266-4316-afb4-e53d95876a76

+++ core/src/java/liquibase/database/sql/DropUniqueConstraintStatement.java
@@ -3,6 +3,7 @@ package liquibase.database.sql;
 import liquibase.database.Database;
 import liquibase.database.MaxDBDatabase;
 import liquibase.database.MySQLDatabase;
+import liquibase.database.SQLiteDatabase;
 import liquibase.exception.StatementNotSupportedOnDatabaseException;
 
 public class DropUniqueConstraintStatement implements SqlStatement {
@@ -30,12 +31,15 @@ public class DropUniqueConstraintStatement implements SqlStatement {
     }
 
     public String getSqlStatement(Database database) throws StatementNotSupportedOnDatabaseException {
-        if (database instanceof MySQLDatabase) {
+    	if (!supportsDatabase(database)) {
+            throw new StatementNotSupportedOnDatabaseException(this, database);
+        }
+    	
+    	if (database instanceof MySQLDatabase) {
             return "ALTER TABLE " + database.escapeTableName(getSchemaName(), getTableName()) + " DROP KEY " + getConstraintName();
         } else if (database instanceof MaxDBDatabase) {
             return "DROP INDEX " + getConstraintName() + " ON " + database.escapeTableName(getSchemaName(), getTableName());
         }
-
 
         return "ALTER TABLE " + database.escapeTableName(getSchemaName(), getTableName()) + " DROP CONSTRAINT " + getConstraintName();
     }
@@ -45,6 +49,6 @@ public class DropUniqueConstraintStatement implements SqlStatement {
     }
 
     public boolean supportsDatabase(Database database) {
-        return true;
+    	return !(database instanceof SQLiteDatabase);
     }
 }
